@@ -3,50 +3,40 @@
     "version": "0.1.0",
     "private": true,
     "scripts": {
-        "start": "ng build --watch --output-hashing none",
+        "serve": "ng run demo-configurator-page:serve",
         "build": "ng build --configuration production --output-hashing none",
-        "test": "ng test",
         "lint": "ng lint"
     },
     "dependencies": {
-        "@angular/animations": "~15.2.9",
-        "@angular/common": "~15.2.9",
-        "@angular/compiler": "~15.2.9",
-        "@angular/core": "~15.2.9",
-        "@angular/forms": "~15.2.9",
-        "@angular/platform-browser": "~15.2.9",
-        "@angular/platform-browser-dynamic": "~15.2.9",
-        "@angular/router": "~15.2.9",
-        "@ngx-translate/core": "~14.0.0",
-        "@ngx-translate/http-loader": "~7.0.0",
-        "rxjs": "~7.5.7",
-        "tslib": "~2.0.0",
-        "zone.js": "~0.12.0"
+        "@angular/animations": "~17.3.1",
+        "@angular/common": "~17.3.1",
+        "@angular/compiler": "~17.3.1",
+        "@angular/core": "~17.3.1",
+        "@angular/forms": "~17.3.1",
+        "@angular/platform-browser": "~17.3.1",
+        "@angular/platform-browser-dynamic": "~17.3.1",
+        "@angular/router": "~17.3.1",
+        "@ngx-translate/core": "~15.0.0",
+        "@ngx-translate/http-loader": "~8.0.0",
+        "rxjs": "~7.8.1",
+        "tslib": "~2.6.2",
+        "zone.js": "~0.14.4"
     },
     "devDependencies": {
-        "@angular-devkit/build-angular": "~15.2.8",
-        "@angular-eslint/builder": "~15.2.1",
-        "@angular-eslint/eslint-plugin": "~15.2.1",
-        "@angular-eslint/eslint-plugin-template": "~15.2.1",
-        "@angular-eslint/schematics": "~15.2.1",
-        "@angular-eslint/template-parser": "~15.2.1",
-        "@angular/cli": "~15.2.8",
-        "@angular/compiler-cli": "~15.2.9",
-        "@types/node": "~14.18.47",
-        "@types/jasmine": "~4.3.1",
-        "@types/jasminewd2": "~2.0.10",
-        "@typescript-eslint/eslint-plugin": "~5.59.7",
-        "@typescript-eslint/parser": "~5.59.7",
-        "eslint": "~8.41.0",
-        "eslint-plugin-deprecation": "~1.4.1",
-        "jasmine-core": "~4.6.0",
-        "karma": "~6.4.2",
-        "karma-chrome-launcher": "~3.2.0",
-        "karma-coverage": "~2.2.0",
-        "karma-coverage-istanbul-reporter": "~3.0.3",
-        "karma-jasmine": "~5.1.0",
-        "karma-jasmine-html-reporter": "~2.0.0",
+        "@angular-devkit/build-angular": "~17.3.2",
+        "@angular-eslint/builder": "~17.3.0",
+        "@angular-eslint/eslint-plugin": "~17.3.0",
+        "@angular-eslint/eslint-plugin-template": "~17.3.0",
+        "@angular-eslint/schematics": "~17.3.0",
+        "@angular-eslint/template-parser": "~17.3.0",
+        "@angular/cli": "~17.3.1",
+        "@angular/compiler-cli": "~17.3.1",
+        "@types/node": "~20.11.30",
+        "@typescript-eslint/eslint-plugin": "7.2.0",
+        "@typescript-eslint/parser": "7.2.0",
+        "eslint": "~8.57.0",
+        "eslint-plugin-deprecation": "~2.0.0",
         "ts-node": "~10.9.1",
-        "typescript": "~4.9.5"
+        "typescript": "~5.4.3"
     }
 }
