@@ -1,8 +1,41 @@
+export interface MockDataItemDisabled {
+    condition: string[];
+    text: string;
+}
+
+export interface MockDataItem {
+    value: string;
+    title: string;
+    price: number;
+    disabled?: MockDataItemDisabled;
+}
+
+export interface MockData {
+    id: string;
+    label: string;
+    tooltip: string;
+    icon: string;
+    data: MockDataItem[];
+}
+
+export interface MockConfigurator {
+    configuration: MockData[];
+    data: {
+        name: string;
+        image: string;
+        logo: string;
+    };
+    defaults: Record<string, string>;
+}
+
+export interface ProductMetaData {
+    timestamp?: string;
+    checkSum?: string;
+}
+
 export interface ProductData {
     sku: string;
     item_group_key: string;
-    id_wishlist_item: number;
-    shopping_list_item_uuid: string;
     quantity: number;
     configurator_key: string;
     customer_reference: string;
@@ -13,8 +46,6 @@ export interface ProductData {
     source_type: string;
     back_url: string;
     submit_url: string;
-    timestamp?: string;
-    checkSum?: string;
 }
 
 export interface ServerData extends ProductData {
@@ -24,25 +55,16 @@ export interface ServerData extends ProductData {
 
 export interface ConfiguredProduct extends ProductData {
     price: number;
-    configuration: DayConfiguration;
-    display_data: DateConfiguration;
+    configuration: Record<string, string>;
+    display_data: Record<string, string>;
     available_quantity: number | string;
-    volume_prices?: VolumePrices | Record<string, never>;
+    volume_prices?: VolumePrices | object;
 }
 
 export interface VolumePrices {
-    volume_prices: {
+    volume_prices: Array<{
         quantity: number;
         net_price: number;
         gross_price: number;
-    }[];
-}
-
-export interface DayConfiguration {
-    time_of_day: number;
-}
-
-export interface DateConfiguration {
-    'Preferred time of the day': string;
-    Date: string;
+    }>;
 }
